refactor + added 404
